+++ documentation/circleci-orbs/orbinoid/orb/src/tests/git-repo/examples/default-server-side/src/create-exposed-app.test.ts
@@ -1,0 +1,7 @@
+import { app } from "./create-exposed-app"
+
+describe("app", () => {
+  test("it should be ok", () => {
+    expect(app.ok).toBe(true)
+  })
+})