+++ documentation/circleci-orbs/orbinoid/orb/src/tests/git-repo/src/create-exposed-app.ts
@@ -1,0 +1,1 @@
+export { createExposedApp } from "./create-exposed-app/create-exposed-app"