+++ documentation/circleci-orbs/orbinoid/orb/src/tests/git-repo/src/create-exposed-app/view-data/npm-author.ts
@@ -1,0 +1,11 @@
+import { AppOptions } from "../options"
+
+export function getNpmAuthor(appOptions: AppOptions): string {
+  return [
+    appOptions.authorName,
+    appOptions.authorEmail && `<${appOptions.authorEmail}>`,
+    appOptions.authorWebsite && `(${appOptions.authorWebsite})`,
+  ]
+    .filter(Boolean)
+    .join(" ")
+}