+++ RELEASE_NOTES.md
@@ -3,3 +3,6 @@
 In this version :
 
 * `Gravitee CI CD Orchestra` `dotenv` defaulting mechanism is fixed
+* `documentation/circleci-orbs/orbinoid/orb` cotnains a much better orb starter,
+* added `orbinoid` integration tests feature,
+* `orbinoid` orb publishing process now also git flow releases the orb source code